Cleaned up testing/fake package
